+++ aalekh-gradle/src/main/kotlin/com/aalekh/aalekh/gradle/AalekhSettingsPlugin.kt
@@ -106,6 +106,9 @@ public class AalekhSettingsPlugin : Plugin<Settings> {
                 task.featurePattern.set(extension.featureIsolationConfig.featurePattern)
                 task.featureAllowedPairs.set(extension.featureIsolationConfig.allowedPairs)
                 task.ruleEntries.set(extension.rulesConfig.entries)
+                task.teamEntries.set(rootProject.provider {
+                    extension.teamOwnership.toInputString()
+                })
                 task.trendFile.set(rootProject.layout.buildDirectory.file("aalekh/trend.json"))
                 task.dependsOn(extractTask)
             }